include the reference porosity in the energy balance
as I understand that matter, the rock heat capacity is specified for
massive rock, i.e. not considering the porosity for the energy equation at
all is also wrong.

(only rock compressibility needs to be disregarded because this would
lead to energy appearing out of thin -- err -- air as the pressure in
the reservoir changes.)