Avoid singul. for zero mole frac. in one-phase
Having a multi-component, single-phase, gravity-driven system with parts
of the domain initialized to zero mole fraction of one of the
components, will result in a singular equation and zero component
fugacity.

This change introduce a small addition to the fugacity, not enough the
skew the results, but enough for the problem to be numerically
tractible.

Credits belong to Tor-Harald, who is the real mastermind behind tracking
down and fixing this problem.