do no longer validate the results of any test
this resolves #469.

since the arguments brought forward in that issue equally apply to the
tests which produce VTK files, and the willingness to manually update
reference results for these is likely even smaller, the results
produced by the eWoms integration tests are simply not validated
anymore besides the fact that the produced files must be non-empty.

Note that this only applies to the default build system; the
alternative system still does -- and will continue to -- validate the
results.