start(): call MPI_Init() after dealing with the command line parameters
some MPI libraries (i.e., openMPI 3.1.1.0, the one installed on my
system) have the annoying habbit of gobbling up about 1 second for
MPI_Init(). While this is not a big deal when running a simulation, it
is really annoying for tab completion. Thus, deal with the parameters
before MPI_Init() so that the help message can be printed
instantaneously.

Note that this assumes that `mpirun` does not mess around with the
command line parameters. This seems to be the case for at least
openMPI and MPICH, but -- as far as I know -- this behavior is not
guaranteed by the MPI standard.

A better approach would be to detect if the simulator is started via
`mpirun` and only defer the call to `MPI_Init()` if this is not the
case. Unfortunately, I am not aware of any way to do this
generically. (It is possible for specific implementations, e.g., for
openMPI the presence of certain environment variables indicate this.)