do not use BCRSMatrix::entry()
this seems to be intended to be only for BCRS matrices which are build
in implicit mode. This might be a mistake in DUNE because the method
works fine for other matrices provided that they are fully build.

This assert is only triggered if the `DUNE_ISTL_WITH_CHECKING` macro
is set, so the impacts of this mistake should be quite limited.