Fixes usage of firstmatrixelement with DUNE 2.8
The function name is now CamelCase and the old
naming scheme is deprecated. Hence we use the new naming
for 2.8
There was also a compilation problem due to header inclusion
order for that function. The compiler did not find our version
the correct version. With this change compilation succeeds again.