ThreadManager: Allow querying number of threads from OpenMP in init.
We actually do set the number of OpenMP threads based on
OMP_NUM_THREADS or command line option ThreadsPerProcess for OpenMP already
in opm-simulators and these settings should just be used here, too. Hence
the new flag for skipping the command line parameter and use the
number of threads that OpenMP has available.

Querying it here results in different behavior as for FlowMainEbos.hpp:
In opm-simulators variable OMP_NUM_THREADS takes precedence over the
command line parameter, but here it would only be used if the command
line option was not used at all.